pyproject.toml
[project]
name = "smartcrm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery>=5.6.3",
    "django>=6.0.3",
    "django-celery-beat>=2.9.0",
    "djangorestframework>=3.17.1",
    "djangorestframework-simplejwt>=5.5.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
]
